Add linter for tests
