+++ dylint_lints/Cargo.toml
@@ -21,11 +21,12 @@ members = [
     "de08_rest_api_conventions/de0803_api_snake_case",
     "de09_gts_layer/de0901_gts_string_pattern",
     "de09_gts_layer/de0902_no_schema_for_on_gts_structs",
+    "de11_testing/de1101_tests_in_separate_files",
     "de03_domain_layer/de0309_must_have_domain_model",
     "de05_client_layer/de0504_client_versioning",
     "de07_security/de0707_drop_zeroize",
     "de13_common_patterns/de1301_no_print_macros",
-    # "de13_common_patterns/de1302_error_from_to_string",
+    "de13_common_patterns/de1302_error_from_to_string",
     "de13_common_patterns/de1303_no_primitive_type_alias",
 ]
 resolver = "3"
@@ -44,6 +45,7 @@ authors = ["Cyber Fabric"]
 # - DE0503-DE0504: Client layer validation (naming conventions, versioning)
 # - DE0801-DE0803: REST API conventions (endpoints, odata, snake_case)
 # - DE0901-DE0902: GTS layer validation (string patterns, schema_for)
+# - DE1101: Resource-group testing policy (tests must be in separate files)
 
 [workspace.lints.rust]
 unsafe_code = "forbid"
