+++ dylint_lints/de11_testing/de1101_tests_in_separate_files/ui/other_module_inline_tests.rs
@@ -1,0 +1,9 @@
+// simulated_dir=/workspace/modules/system/types-registry/types-registry/src/api/rest/
+#[cfg(test)]
+// Should not trigger DE1101 - tests must be in separate files
+mod tests {
+    #[test]
+    fn inline_test() {}
+}
+
+fn main() {}
