+++ dylint_lints/de11_testing/de1101_tests_in_separate_files/ui/resource_group_inline_case.rs
@@ -1,0 +1,9 @@
+// simulated_dir=/workspace/modules/system/resource-group/resource-group/src/api/rest/
+#[cfg(test)]
+// Should trigger DE1101 - tests must be in separate files
+mod tests {
+    #[test]
+    fn inline_test() {}
+}
+
+fn main() {}
