+++ dylint_lints/de11_testing/de1101_tests_in_separate_files/ui/resource_group_out_of_line_test_module.rs
@@ -1,0 +1,7 @@
+// simulated_dir=/workspace/modules/system/resource-group/resource-group/src/api/rest/
+#[cfg(test)]
+#[path = "dto_tests.rs"]
+// Should not trigger DE1101 - tests must be in separate files
+mod tests;
+
+fn main() {}
