+++ dylint_lints/de11_testing/de1101_tests_in_separate_files/ui/resource_group_separate_test_file_tests.rs
@@ -1,0 +1,6 @@
+// simulated_dir=/workspace/modules/system/resource-group/resource-group/src/api/rest/
+// Should not trigger DE1101 - tests must be in separate files
+#[test]
+fn separate_test_file_is_allowed() {}
+
+fn main() {}
